Fix: 변경 회원정보수정, 회원탈퇴

See merge request s06-webmobile1-sub2/S06P12A201!59

+++ backend/src/main/java/com/ssafy/api/controller/UserController.java
@@ -168,7 +168,9 @@ public class UserController {
         @ApiResponse(code = 400, message = "존재하지 않는 유저입니다."),
         @ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<? extends BaseResponseBody> deleteUser(@PathVariable("userId") int userId) {
+	public ResponseEntity<? extends BaseResponseBody> deleteUser(@PathVariable("userId") int userId, String password) {
+		if(!passwordEncoder.matches(password, userService.getUserById(userId).getPassword())) // 패스워드 일치 확인
+			return ResponseEntity.status(400).body(BaseResponseBody.of(400, "잘못된 비밀번호"));
 		userService.deleteUser(userId);
 		return ResponseEntity.status(200).body(BaseResponseBody.of(200, "유저 탈퇴 성공"));
 	}
